Rename WeekContainer -> CalendarGrid, deprecate Week component
