+++ src/client/components/Pages/Auth/Contributions/CalendarGrid/DayCell/DayCell.tsx
@@ -25,7 +25,6 @@ export const DayCell: React.FC<DayCellProps> = (props) => {
         <div
             style={{
                 position: 'relative',
-                margin: 1,
             }}
         >
             <div
