+++ src/client/components/Pages/Auth/Contributions/CalendarGrid/index.ts
@@ -1,0 +1,2 @@
+import { CalendarGrid } from './CalendarGrid'
+export default CalendarGrid
