--- src/client/components/Pages/Auth/Contributions/WeekContainer/Week/Week.module.css
@@ -1,5 +1,0 @@
-.contribution_week {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-}
