--- src/client/components/Pages/Auth/Contributions/WeekContainer/Week/index.ts
@@ -1,2 +1,0 @@
-import { Week } from './Week'
-export default Week
