--- src/client/components/Pages/Auth/Contributions/WeekContainer/index.ts
@@ -1,2 +1,0 @@
-import { WeekContainer } from './WeekContainer'
-export default WeekContainer
